Eric Sunshine eliminated compilation warnings throughout project.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18711 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/engine/3d/engine.cpp
@@ -1609,7 +1609,8 @@ void csEngine::ShineLights (iRegion *region, iProgressMeter *meter)
   }
   if (failed > 0)
   {
-    Warn ("Couldn't load cached lighting for %d object(s):", failed);
+    Warn ("Couldn't load cached lighting for %lu object(s):",
+	  (unsigned long)failed);
     size_t i;
     for (i = 0 ; i < failed_meshes.Length () ; i++)
     {
